Organized the codebase properly
